Takes the Query String and searches the String for matches.
Takes the Query String and searches the String for matches returns the index position of the word located inside str. This program is case sensitive so if your string is "helloWorld" and your query string is "world" the program would return -1. To ignore case, before comparing the temporary array and the queryStr in the if statement. add tempArr.toLowerCase() and querystr.toLowerCase. Since both are now changed to lowercase, case will be ignored  and as long as the words match, the program should return an index value.